refactor(aws): rename dynamodb.provider.ts to dynamodb.ts
