+++ aws/src/providers/dynamodb.ts
